Changed the colors of the dark theme to more pleasant ones
I couldn't change the color of the svg icons. Because of this, they do not look very nice on a dark background. If you change their color to a lighter one (It is desirable that the icons on the gray colors be #b9bbbe, and on the colors of the purple and green buttons - #ffffff), then everything should look beautiful.